Refactor AnnotationValueExpressionUtil
Change direct package use to import
Update test cases for quick fixes
